pull-split-exceptions.py (#48)
* moved all exceptions to relevant submodules

* as they are not tested - refactor to _name*.py and done
