+++ src/landiv_blur/_exceptions.py
@@ -1,0 +1,1 @@
+# TODO: delete once ready
